Derive Raft Id from PeerId
Instead of requiring users to come up with and specify a mapping for
each validator's public key to Raft Id, the Raft Id is derived directly
from the public key.

For the size of networks typically used in a Raft deployment (and even
much larger networks) the chance of collision should be small enough
that it outweighs the cost of manually configuring this.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>